feat: scenes are removed. Improving mechanism to render something

+++ src/main/kotlin/com/conceptic/firefly/screen/support/MouseActionsPublisher.kt
@@ -3,7 +3,7 @@ package com.conceptic.firefly.screen.support
 import com.conceptic.firefly.support.Publisher
 
 interface MouseActionsSubscriber {
-    fun onMouseMove(mouseX: Int, mouseY: Float)
+    fun onMouseMove(mouseX: Float, mouseY: Float)
 
     fun onRightClick()
